+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11MethodParser.java
@@ -1,12 +1,10 @@
 package org.apache.coyote.http11.request;
 
-import org.apache.coyote.http11.Http11StartLineParser;
-
 class Http11MethodParser {
 
     private final Http11StartLineParser startLineParser;
 
-    public Http11MethodParser(Http11StartLineParser startLineParser) {
+    Http11MethodParser(Http11StartLineParser startLineParser) {
         this.startLineParser = startLineParser;
     }
