+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11RequestUriParser.java
@@ -1,8 +1,6 @@
 package org.apache.coyote.http11.request;
 
-import org.apache.coyote.http11.Http11StartLineParser;
-
-public class Http11RequestUriParser {
+class Http11RequestUriParser {
 
     private final Http11StartLineParser startLineParser;
 
@@ -13,7 +11,7 @@ public class Http11RequestUriParser {
     /*
             https://www.rfc-editor.org/rfc/rfc2616#section-5 이 문서에서 이 메서드가 반환하는 것을 Request-URI라 지칭합니다.
         */
-    public String parseRequestURI(String requestMessage) {
+    String parseRequestURI(String requestMessage) {
         String startLine = startLineParser.parseStartLine(requestMessage);
         return startLine.split(" ")[1];
     }
